+++ App/include/v8world/Poly.h
@@ -53,9 +53,9 @@ namespace RBX {
 		virtual CoordinateFrame getSurfaceCoordInBody( const size_t surfaceId ) const;
 		Vector3 getSurfaceNormalInBody( const size_t surfaceId ) const;
 		size_t getMostAlignedSurface( const Vector3& vecInWorld, const G3D::Matrix3& objectR ) const;
-		int getNumSurfaces( void ) const { return mesh->numFaces(); }
-		Vector3 getSurfaceVertInBody( const size_t surfaceId, const int vertId ) const;
-		int getNumVertsInSurface( const size_t surfaceId ) const;
+		size_t getNumSurfaces( void ) const { return mesh->numFaces(); }
+		Vector3 getSurfaceVertInBody( const size_t surfaceId, const size_t vertId ) const;
+		size_t getNumVertsInSurface( const size_t surfaceId ) const;
 		bool vertOverlapsFace( const Vector3& pointInBody, const size_t surfaceId ) const;
 
         /*override*/virtual bool findTouchingSurfacesConvex( const CoordinateFrame& myCf, size_t& myFaceId, const Geometry& otherGeom, const CoordinateFrame& otherCf, size_t& otherFaceId ) const;
